env sample

+++ server/src/controllers/journalController.js
@@ -1,4 +1,5 @@
 import Journal from "../models/JournalModel.js";
+import User from "../models/UserModel.js";
 import { v4 as uuidv4 } from "uuid";
 import cloudinary from "cloudinary";
 import { asyncHandler } from "../middlewares/asyncHandler.js";
@@ -7,27 +8,43 @@ import ErrorHandler from "../utils/ErrorHandler.js";
 class JournalController {
   static createJournal = asyncHandler(async (req, res, next) => {
     try {
-      const { title, content, mood, location, tags, images, isPrivate } =
-        req.body;
-      const user = req.user;
+      const {
+        title,
+        content,
+        mood,
+        location,
+        tags,
+        images,
+        isPrivate = true,
+      } = req.body;
 
       if (!title && !content && !tags && !images) {
-        return next(
-          new ErrorHandler("Title, content, tags and images are required", 404)
-        );
+        let errors = {
+          title: "Title is required",
+          content: "Content is required",
+          tags: "At least one Tag is required",
+          images: "At least one image is required",
+        };
+        return next(new ErrorHandler("Journal Creation Failed!", 404, errors));
       }
 
       // Handle Image section using cloudinary
 
+      // Increment the user Journal Entry
+      await User.findByIdAndUpdate(req.user._id, {
+        $inc: { journalEntries: 1 },
+      });
+
       // Create a new Journal
       let journal = await Journal.create({
-        author: user._id,
+        author: req.user._id,
         title,
         content,
         mood,
         location,
         tags,
         isPrivate,
+        images,
       });
 
       return res.status(201).json({ success: true, journal });
@@ -87,11 +104,26 @@ class JournalController {
   static deleteJournal = asyncHandler(async (req, res, next) => {
     try {
       const id = req.params.id;
+
+      const author = await User.findById(req.user._id);
+
+      if (!author) {
+        return res
+          .status(404)
+          .json({ success: false, message: "Author not found!" });
+      }
+
       const deletedJournal = await Journal.findByIdAndDelete(id);
 
       if (!deletedJournal) {
         return next(new ErrorHandler("Journal was not found", 400));
       }
+
+      // Decrement the user Journal Entries
+      await User.findByIdAndUpdate(req.user._id, {
+        $inc: { journalEntries: -1 },
+      });
+
       return res
         .status(200)
         .json({ status: true, message: "Journal was deleted Successfully" });
